Cleaned up part 2
I like programming recursively, but when I initially tried to solve part 2 this way I was missing a few concepts. I wasn't thinking that about how the way you face the block doesn't matter; if you've previously checked if a blocker doesn't work at a coordinate, it doesn't matter if you reach that point in the future; you've already proven it doesn't work. If you block it _later_, you'll have been blocking yourself earlier on. I didn't think of this at first, and so re-wrote my part 1 to not be recursive (maybe should have done this anyway, since I'd had to increase the recursion depth). Took a while to run part 2, but I don't think it's _quite_ brute force. Not the most efficient algorithm though.

I'd had to take a few breaks solving this part because the wife was getting upset I was coding for so long. May not be able to spend much time coding at home anymore... Life doesn't pause, even if you're _really close_ to solving the problem.